requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.2
djangorestframework==3.12.2
pip==20.2.4
psycopg2==2.8.5
pytz==2020.1
setuptools==50.3.0
sqlparse==0.4.1
wheel==0.35.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib ==0.2.2
six==1.15.0

